CI: initial preparations AppImage compiling
Added initial steps in CI/CD pipeline for compiling AppImages, to remove the necessity of installing .NET Framework separately on a Linux machine - allowing systems which can't install the framework because of an immutable filesystem (e.g. Fedora Silverblue, SteamOS 3) to run Mesen. This is planned to be done through Kuiper Zone's PupNet Deploy tool. The usage of this tool could get extended to also package into the Flatpak package format for a similar result.

Left to-do (as currently known):

- Add .conf file for configuration of PupNet Deploy builds
- Add any other files necessary for AppImage compilation
- Test the added pipeline commands in a Linux environment (WSL2)