SNES: Add support for ExLoRom mapping
This format is used by ROM hacks, and is supported by at least Snes9x
and bsnes. It's indicated by having a copy of the header at 0x400000
with a LoRom map mode.

The removed check for map mode 0x22 triggered in S-DD1 games, but their
memory layout is very different from ExLoRom. Their mapping is handled
by `MapBsxMemoryPack` instead of the regular `RegisterHandlers` code
anyway, so the setting of that flag had no effect other than what's
reported in the log window, but it was removed to prevent future
confusion.